Fix : 백엔드와의 통신 완료
