--- src/pages/Product/Components/ProductDetail/ProductInfo.scss
@@ -1,246 +1,0 @@
-@import '../../../../styles/variable.scss';
-@mixin fontStyle {
-  color: #333;
-  font-size: 17px;
-  line-height: 1.4;
-  letter-spacing: -0.5px;
-}
-
-.productInfo {
-  display: flex;
-  width: 90%;
-  height: 600px;
-  margin-top: 100px;
-  padding-top: 50px;
-
-  .productThumbnail {
-    @include flex-center;
-    width: 45%;
-  }
-
-  .itemInfomation {
-    width: 55%;
-    padding: 10px 10px 10px 40px;
-
-    .navigatorWrap {
-      text-align: right;
-      .whereIsIt {
-        font-size: 17px;
-        color: #333;
-      }
-    }
-
-    .itemTitle {
-      padding-bottom: 20px;
-      margin-bottom: 20px;
-      border-bottom: 1px solid #e7e7e7;
-
-      .itemName {
-        padding-right: 60px;
-        line-height: 34px;
-        font-weight: normal;
-        font-size: 40px;
-        font-weight: bold;
-        font-family: 'notokrM';
-      }
-
-      .itemTag {
-        padding-top: 18px;
-        font-size: 18px;
-        color: #808080;
-      }
-    }
-
-    .itemInfo {
-      @include fontStyle();
-      position: relative;
-      margin-bottom: 20px;
-      padding-bottom: 20px;
-      border-bottom: 1px solid #e7e7e7;
-
-      .markON {
-        position: absolute;
-        top: 14px;
-        width: 500px;
-        height: 260px;
-        padding: 12px;
-        border: 1px solid #808080;
-        background-color: #f3f3f3;
-        z-index: 1;
-      }
-
-      .markOFF {
-        display: none;
-      }
-
-      .goodToKnow {
-        .questionMark {
-          width: 18px;
-          margin-left: 5px;
-        }
-      }
-
-      .itemInfoList {
-        margin-top: 50px;
-
-        .itemInfoRow {
-          margin: 15px 0;
-
-          .itemInfoDetail {
-            color: #808080;
-            margin-right: 30px;
-          }
-
-          .itemCountBox {
-            display: inline-block;
-            height: 100px;
-
-            .countMinus,
-            .countPlus {
-              width: 40px;
-              height: 40px;
-              border: 1px solid #808080;
-              font-weight: bold;
-              background-color: white;
-            }
-
-            .productCount {
-              width: 80px;
-              height: 40px;
-              font-weight: bold;
-              text-align: center;
-              border: 1px solid #808080;
-              border-right: 0;
-              border-left: 0;
-            }
-          }
-        }
-      }
-    }
-
-    .itemPrice {
-      margin: 30px 0;
-      text-align: right;
-
-      .howMuch {
-        margin-right: 20px;
-        color: #333;
-      }
-
-      .allPrice {
-        font-size: 23pt;
-        font-weight: bold;
-      }
-    }
-
-    .btnList {
-      display: flex;
-      justify-content: space-between;
-
-      .cartBtn {
-        width: 60%;
-        height: 50px;
-        margin-right: 10px;
-        outline: 0;
-        border: 1px solid black;
-        font-size: 12pt;
-        background-color: #fff;
-      }
-
-      .buyBtn {
-        width: 60%;
-        height: 50px;
-        outline: 0;
-        border: 0;
-        font-size: 12pt;
-        color: #fff;
-        background-color: black;
-      }
-    }
-  }
-}
-
-.photoReviewWrap {
-  width: 90%;
-
-  .photoReviewText {
-    font-size: 20pt;
-    font-weight: bold;
-  }
-
-  .photoList {
-    display: flex;
-    justify-content: space-between;
-    margin-top: 20px;
-
-    .photoListWrap {
-      width: 64%;
-      padding: 20px;
-      border: 1px solid #808080;
-
-      .photoListText {
-        margin-bottom: 20px;
-        font-size: 13pt;
-        color: #808080;
-      }
-
-      .photoImgList {
-        .photoWrap {
-          .reviewImg {
-            width: 100px;
-            height: 100px;
-            margin-right: 10px;
-
-            &:last-child {
-              margin-right: 0px;
-            }
-          }
-        }
-      }
-    }
-
-    .reviewCountWrap {
-      display: flex;
-      width: 34%;
-      padding: 20px;
-      border: 1px solid #808080;
-
-      .productScore {
-        width: 50%;
-        padding-top: 25px;
-        text-align: center;
-
-        .scoreText {
-          font-size: 16pt;
-          margin-bottom: 10px;
-        }
-
-        .score {
-          font-size: 35pt;
-          font-weight: bold;
-          margin-bottom: 10px;
-        }
-
-        .scoreStars {
-          font-size: 16pt;
-        }
-      }
-
-      .reviewCounter {
-        width: 50%;
-        padding-top: 40px;
-        text-align: center;
-
-        .reviewCountInfo {
-          font-size: 16pt;
-          margin-bottom: 20px;
-
-          .infoName {
-            color: #808080;
-            margin-right: 10px;
-          }
-        }
-      }
-    }
-  }
-}
