--- src/pages/Product/Components/ProductDetail/ProductIntroduce.scss
@@ -1,26 +1,0 @@
-@import '../../../../styles/variable.scss';
-
-.itemIntroduce {
-  @include flex-center();
-  flex-direction: column;
-  width: 90%;
-
-  .itemInfoImg {
-    width: 100%;
-  }
-
-  .textArea {
-    padding: 40px;
-    font-family: 'notokr', sans-serif;
-    font-size: 20px;
-    font-weight: 400;
-
-    .how {
-      display: inline-block;
-      margin-bottom: 10px;
-      color: white;
-      background-color: black;
-      font-size: 40pt;
-    }
-  }
-}
